refactor: move api types to an independent module
