+++ src/components/GraphChannelList/index.tsx
@@ -3,8 +3,8 @@ import { Tooltip } from 'antd'
 import dayjs from 'dayjs'
 import type { FC } from 'react'
 import { Link } from 'react-router-dom'
+import type { Fiber } from '../../services/ExplorerService'
 import { TIME_TEMPLATE } from '../../constants/common'
-import type { Fiber } from '../../services/ExplorerService/fetcher'
 import { parseNumericAbbr } from '../../utils/chart'
 import { formalizeChannelAsset } from '../../utils/fiber'
 import { localeNumberString } from '../../utils/number'
