+++ src/pages/Fiber/GraphNodeList/index.tsx
@@ -7,7 +7,7 @@ import dayjs from 'dayjs'
 import Content from '../../../components/Content'
 import { useSetToast } from '../../../components/Toast'
 import { explorerService } from '../../../services/ExplorerService'
-import type { Fiber } from '../../../services/ExplorerService/fetcher'
+import type { Fiber } from '../../../services/ExplorerService'
 import { ReactComponent as CopyIcon } from '../../../components/Copy/icon.svg'
 import Pagination from '../Pagination'
 import { PAGE_SIZE, TIME_TEMPLATE } from '../../../constants/common'
