+++ src/pages/NftCollectionInfo/NftCollectionTransfers/index.tsx
@@ -5,11 +5,10 @@ import { useTranslation } from 'react-i18next'
 import { Link } from '../../../components/Link'
 import { getImgFromSporeCell } from '../../../utils/spore'
 // TODO: Refactor is needed. Should not directly import anything from the descendants of ExplorerService.
-import type { TransferListRes, TransferRes } from '../../../services/ExplorerService/fetcher'
 import styles from './styles.module.scss'
 import { getPrimaryColor } from '../../../constants/common'
 import { formatNftDisplayId, handleNftImgError, hexToBase64, patchMibaoImg } from '../../../utils/util'
-import { explorerService } from '../../../services/ExplorerService'
+import { type TransferListRes, TransferRes, explorerService } from '../../../services/ExplorerService'
 import { dayjs } from '../../../utils/date'
 import { useParsedDate, useTimestamp } from '../../../hooks'
 import { useCurrentLanguage } from '../../../utils/i18n'
