+++ src/pages/NftCollections/List.tsx
@@ -3,6 +3,7 @@ import { Popover, Tooltip } from 'antd'
 import classNames from 'classnames'
 import { Trans, useTranslation } from 'react-i18next'
 import { TFunction } from 'i18next'
+import type { NFTCollection } from '../../services/ExplorerService'
 import { Link } from '../../components/Link'
 import SortButton from '../../components/SortButton'
 import { handleNftImgError, patchMibaoImg } from '../../utils/util'
@@ -11,7 +12,6 @@ import { ReactComponent as FilterIcon } from './filter.svg'
 import { getPrimaryColor } from '../../constants/common'
 import { useIsMobile, useSearchParams } from '../../hooks'
 import styles from './styles.module.scss'
-import type { NFTCollection } from '../../services/ExplorerService/fetcher'
 import { useNFTCollectionsSortParam } from './util'
 import { parseSimpleDate } from '../../utils/date'
 import MultiFilterButton from '../../components/MultiFilterButton'
