+++ src/pages/Script/index.tsx
@@ -15,8 +15,7 @@ import { usePaginationParamsInPage } from '../../hooks'
 import { shannonToCkb } from '../../utils/util'
 import Capacity from '../../components/Capacity'
 import styles from './styles.module.scss'
-import { explorerService } from '../../services/ExplorerService'
-import type { ScriptInfo } from '../../services/ExplorerService/fetcher'
+import { type ScriptInfo, explorerService } from '../../services/ExplorerService'
 import { ScriptTab, ScriptTabPane, ScriptTabTitle } from './styled'
 import { Card, CardCellInfo, CardCellsLayout } from '../../components/Card'
 import { ReactComponent as OpenSourceIcon } from '../../assets/open-source.svg'
